Bringing build/test/lint setup up to our modern standards. Will fix lint in a later PR for clarity.
