--- .eslintrc
@@ -1,186 +1,0 @@
----
-globals:
-  # We're just going to assume a CommonJS / Webpack world...
-  require: false
-  module: false
-  process: false
-
-ecmaFeatures:
-  jsx: true
-  modules: true
-
-env:
-  es6: true
-  node: true
-  browser: true
-
-globals:
-  fetch: false                      # Expect a `fetch` polyfill.
-  CodeMirror: false
-
-rules:
-  no-extra-parens: 0  # Ugh. Want `return (<Foo />);` to be allowed.
-  no-var: 2
-  no-alert: 2
-  no-array-constructor: 2
-  no-bitwise: 2
-  no-caller: 2
-  no-catch-shadow: 2
-  no-comma-dangle: 2
-  no-cond-assign: 2
-  no-console: 2
-  no-constant-condition: 2
-  no-continue: 0
-  no-control-regex: 2
-  no-debugger: 2
-  no-delete-var: 2
-  no-div-regex: 0
-  no-dupe-keys: 2
-  no-dupe-args: 2
-  no-duplicate-case: 2
-  no-else-return: 0
-  no-empty: 2
-  no-empty-class: 2
-  no-empty-label: 2
-  no-eq-null: 0
-  no-eval: 2
-  no-ex-assign: 2
-  no-extend-native: 2
-  no-extra-bind: 2
-  no-extra-boolean-cast: 2
-  no-extra-semi: 2
-  no-extra-strict: 2
-  no-fallthrough: 2
-  no-floating-decimal: 0
-  no-func-assign: 2
-  no-implied-eval: 2
-  no-inline-comments: 0
-  no-inner-declarations: [2, "functions"]
-  no-invalid-regexp: 2
-  no-irregular-whitespace: 2
-  no-iterator: 2
-  no-label-var: 2
-  no-labels: 2
-  no-lone-blocks: 2
-  no-lonely-if: 2
-  no-loop-func: 2
-  no-mixed-requires: 2
-  no-mixed-spaces-and-tabs: 2
-  no-multi-spaces: 2
-  no-multi-str: 2
-  no-multiple-empty-lines: [2, { max: 2 }]
-  no-native-reassign: 2
-  no-negated-in-lhs: 2
-  no-nested-ternary: 2
-  no-new: 2
-  no-new-func: 2
-  no-new-object: 2
-  no-new-require: 2
-  no-new-wrappers: 2
-  no-obj-calls: 2
-  no-octal: 2
-  no-octal-escape: 2
-  no-param-reassign: 0
-  no-path-concat: 0
-  no-plusplus: 0
-  no-process-env: 0
-  no-process-exit: 2
-  no-proto: 2
-  no-redeclare: 2
-  no-regex-spaces: 2
-  no-reserved-keys: 0
-  no-restricted-modules: 0
-  no-return-assign: 2
-  no-script-url: 2
-  no-self-compare: 2
-  no-sequences: 2
-  no-shadow: 2
-  no-shadow-restricted-names: 2
-  no-space-before-semi: 0
-  no-spaced-func: 2
-  no-sparse-arrays: 2
-  no-sync: 0
-  no-ternary: 0
-  no-trailing-spaces: 2
-  no-throw-literal: 0
-  no-undef: 2
-  no-undef-init: 2
-  no-undefined: 0
-  no-underscore-dangle: 0
-  no-unreachable: 2
-  no-unused-expressions: 2
-  no-unused-vars: [2, {
-    vars: "all",
-    args: "after-used" }]
-  no-use-before-define: 2
-  no-void: 0
-  no-var: 0
-  no-warning-comments: 0
-  no-with: 2
-  no-wrap-func: 2
-  block-scoped-var: 0
-  brace-style: [2, "1tbs", { "allowSingleLine": true }]
-  camelcase: 2
-  comma-dangle: [2, "never"]
-  comma-spacing: 2
-  comma-style: [2, "last"]
-  complexity: [2, 11]
-  consistent-return: 2
-  consistent-this: [2, "self"]
-  curly: [2, "all"]
-  default-case: 0
-  dot-notation: [2, { allowKeywords: true }]
-  eol-last: 2
-  eqeqeq: 2
-  func-names: 0
-  func-style: [2, "expression"]
-  generator-star: 0
-  generator-star-spacing: 0
-  global-strict: [2, "always"]
-  guard-for-in: 0
-  handle-callback-err: 0
-  indent: [2, 2]
-  key-spacing: [2, {
-    beforeColon: false,
-    afterColon: true }]
-  max-depth: [2, 4]
-  max-len: [2, 100, 2]
-  max-nested-callbacks: [2, 3]
-  max-params: [2, 3]
-  max-statements: [2, 15]
-  new-cap: 2
-  new-parens: 2
-  newline-after-var: 0
-  object-shorthand: 0
-  one-var: [2, "never"]
-  operator-assignment: [2, "always"]
-  operator-linebreak: 0
-  padded-blocks: 0
-  quote-props: 0
-  quotes: [2, "double"]
-  radix: 0
-  semi: 2
-  semi-spacing: [2, { before : false, after: true}]
-  sort-vars: 0
-  space-after-keywords: [2, "always"]
-  space-before-blocks: [2, "always"]
-  space-before-function-paren: [2, {
-    anonymous: "always",
-    named: "never" }]
-  space-in-brackets: [0, "always"]
-  space-in-parens: [2, "never"]
-  space-infix-ops: 2
-  space-return-throw-case: 2
-  space-unary-ops: [2, {
-    words: true,
-    nonwords: false }]
-  spaced-line-comment: [2, "always"]
-
-  strict: 0
-  use-isnan: 2
-  valid-jsdoc: 2
-  valid-typeof: 2
-  vars-on-top: 0
-  wrap-iife: 0
-  wrap-regex: 0
-  yoda: [2, "never"]
