+++ .eslintrc-base
@@ -1,0 +1,6 @@
+---
+plugins:
+  - "filenames"
+
+rules:
+  "filenames/filenames": [2, "^[a-z0-9\\-\\.]+$"] # dash-cased filenames.
