+++ .eslintrc-node
@@ -1,0 +1,4 @@
+---
+extends:
+  - "defaults/configurations/walmart/es5-node"
+  - ".eslintrc-base"
