+++ .eslintrc-react
@@ -1,0 +1,4 @@
+---
+extends:
+  - "defaults/configurations/walmart/es6-react"
+  - ".eslintrc-base"
