+++ .eslintrc-react-test
@@ -1,0 +1,8 @@
+---
+extends:
+  - "defaults/configurations/walmart/es6-react"
+  - "defaults/environments/test"
+  - ".eslintrc-base"
+
+globals:
+  expect: false
