+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+Contributing
+============
+
+Thanks for helping out!
+
+## Development
+
+Run `npm run demo` to run a webpack dev server with component examples.
+
+## Checks, Tests
+
+Run `npm run check` before committing.
+
+## Dist
+
+Please do not commit changes to files in `dist`.
+These files are only committed when we tag releases.
