+++ demo/index.html
@@ -11,11 +11,16 @@
   <!--[if lt IE 8]>
     <p class="browsehappy">You are using an <strong>outdated</strong> browser. Please <a href="http://browsehappy.com/">upgrade your browser</a> to improve your experience.</p>
   <![endif]-->
-  <div id="root">
-    <!-- this is where the root react component will get rendered -->
+  <div id="content">
+    <h1>If you can see this, something is broken (or JS is not enabled)!!.</h1>
   </div>
   <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/codemirror/5.0.0/codemirror.min.js"></script>
   <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/codemirror/5.0.0/mode/javascript/javascript.min.js"></script>
-  <script type="text/javascript" src="/demo/assets/bundle.js"></script>
+  <script type="text-javascript" src="http://localhost:3000/webpack-dev-server.js"></script>
+  <script type="text/javascript" src="assets/main.js"></script>
+  <!--[if lt IE 9]>
+    <script src="//cdnjs.cloudflare.com/ajax/libs/es5-shim/4.1.7/es5-shim.min.js"></script>
+    <script src="//cdnjs.cloudflare.com/ajax/libs/es5-shim/4.1.7/es5-sham.min.js"></script>
+  <![endif]-->
 </body>
 </html>
