--- index.js
@@ -1,5 +1,0 @@
-"use strict";
-
-var Playground = require("./lib/playground");
-
-module.exports = Playground;
