+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "component-playground",
+  "version": "0.1.1",
   "description": "A component for rendering React components with editable source and live preview",
   "keywords": [
     "React",
@@ -7,67 +8,97 @@
     "live",
     "component"
   ],
-
-  "version": "0.1.1",
-  "author": "Ken Wheeler",
-
-  "homepage": "https://github.com/FormidableLabs/component-playground#readme",
-  "bugs": {
-    "url": "https://github.com/FormidableLabs/component-playground/issues"
-  },
-
+  "main": "lib/index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/FormidableLabs/component-playground.git"
   },
-  "private": false,
-
+  "author": "Ken Wheeler",
+  "license": "MIT",
+  "homepage": "https://github.com/FormidableLabs/component-playground#readme",
+  "bugs": {
+    "url": "https://github.com/FormidableLabs/component-playground/issues"
+  },
+  "scripts": {
+    "postinstall": "npm run build-lib",
+    "preversion": "npm run check",
+    "version": "npm run clean && npm run build && git add -A dist",
+    "clean-dist": "rimraf dist",
+    "build-dist-min": "webpack --config webpack.config.js",
+    "build-dist-dev": "webpack --config webpack.config.dev.js",
+    "build-dist": "npm run clean-dist && npm run build-dist-min && npm run build-dist-dev",
+    "clean-lib": "rimraf lib",
+    "build-lib": "npm run clean-lib && babel --stage 0 src -d lib",
+    "clean": "npm run clean-lib && npm run clean-dist",
+    "build": "npm run build-lib && npm run build-dist",
+    "server-dev": "webpack-dev-server --port 3000 --config demo/webpack.config.dev.js --content-base demo",
+    "server-hot": "webpack-dev-server --port 3000 --config demo/webpack.config.hot.js --hot --content-base demo",
+    "server-test": "webpack-dev-server --port 3001 --config webpack.config.test.js",
+    "dev": "npm run server-dev & npm run server-test",
+    "hot": "npm run server-hot & npm run server-test",
+    "open-demo": "opener http://127.0.0.1:3000",
+    "open-dev": "npm run dev & npm run open-demo",
+    "open-hot": "npm run hot & npm run open-demo",
+    "lint-node": "eslint -c .eslintrc-node *.js demo/webpack.*.js",
+    "lint-react": "eslint --ext .js,.jsx -c .eslintrc-react src demo/*.jsx",
+    "lint-react-test": "eslint --ext .js,.jsx -c .eslintrc-react-test src test/client",
+    "lint": "npm run lint-node && npm run lint-react && npm run lint-react-test",
+    "test-frontend": "node node_modules/karma/bin/karma start karma.conf.js",
+    "test-frontend-ci": "node node_modules/karma/bin/karma start --browsers PhantomJS,Firefox karma.conf.coverage.js",
+    "test-frontend-cov": "node node_modules/karma/bin/karma start karma.conf.coverage.js",
+    "test-frontend-dev": "node node_modules/karma/bin/karma start karma.conf.dev.js",
+    "test": "npm run test-frontend && echo 'TODO Server 13, Integration 12 tests.'",
+    "test-ci": "npm run test-frontend-ci",
+    "test-cov": "npm run test-frontend-cov",
+    "test-dev": "npm run test-frontend-dev",
+    "check": "npm run lint && npm run test",
+    "check-ci": "npm run lint && npm run test-ci",
+    "check-cov": "npm run lint && npm run test-cov",
+    "check-dev": "npm run lint && npm run test-dev"
+  },
   "dependencies": {
-    "babel": "^5.8.23",
-    "babel-core": "^5.8.24"
+    "babel": "^5.5.8",
+    "babel-core": "^5.5.8",
+    "babel-loader": "^5.3.2",
+    "css-loader": "~0.9.0",
+    "rimraf": "^2.4.0",
+    "style-loader": "~0.8.0",
+    "url-loader": "~0.5.5",
+    "webpack": "^1.10.0"
   },
   "peerDependencies": {
     "react": "0.13.x"
   },
   "devDependencies": {
-    "babel-eslint": "^3.1.20",
-    "babel-loader": "^5.3.0",
-    "chai": "^3.0.0",
-    "css-loader": "~0.15.1",
-    "del": "^2.0.2",
-    "gulp": "^3.8.11",
-    "gulp-babel": "^5.1.0",
-    "gulp-eslint": "^0.15.0",
-    "gulp-open": "^1.0.0",
-    "gulp-util": "^3.0.4",
-    "gulp-webpack": "^1.3.0",
-    "istanbul": "^0.3.13",
-    "istanbul-instrumenter-loader": "^0.1.2",
-    "karma": "~0.13.9",
-    "karma-chrome-launcher": "~0.2.0",
-    "karma-coverage": "^0.5.2",
-    "karma-firefox-launcher": "~0.1.3",
+    "babel-eslint": "^3.1.25",
+    "chai": "^3.2.0",
+    "codemirror": "^5.6.0",
+    "eslint": "^0.24.1",
+    "eslint-config-defaults": "^3.0.3",
+    "eslint-plugin-filenames": "^0.1.1",
+    "eslint-plugin-react": "^2.6.4",
+    "isparta-loader": "^0.2.0",
+    "karma": "^0.12.37",
+    "karma-chrome-launcher": "^0.2.0",
+    "karma-coverage": "^0.4.2",
+    "karma-firefox-launcher": "^0.1.6",
+    "karma-ie-launcher": "^0.2.0",
     "karma-mocha": "^0.2.0",
-    "karma-mocha-reporter": "^1.0.2",
-    "karma-phantomjs-launcher": "~0.2.0",
-    "karma-script-launcher": "~0.1.0",
-    "karma-sinon-chai": "^1.0.0",
-    "karma-webpack": "^1.5.0",
-    "mocha": "^2.2.4",
+    "karma-phantomjs-launcher": "^0.2.0",
+    "karma-phantomjs-shim": "^1.0.0",
+    "karma-safari-launcher": "^0.1.1",
+    "karma-sauce-launcher": "^0.2.14",
+    "karma-spec-reporter": "0.0.20",
+    "karma-webpack": "^1.6.0",
+    "lodash": "^3.9.3",
+    "mocha": "^2.2.5",
+    "opener": "^1.4.1",
+    "phantomjs": "^1.9.17",
     "raw-loader": "^0.5.1",
-    "react": "^0.13.1",
-    "sinon": "^1.14.1",
-    "sinon-chai": "^2.7.0",
-    "style-loader": "~0.12.3",
-    "webpack": "^1.10.1",
-    "webpack-dev-server": "^1.7.0"
-  },
-
-  "main": "index.js",
-  "engines":{
-    "node": ">= 0.10.0"
-  },
-  "scripts": {},
-
-  "license": "MIT"
+    "react": "0.13.x",
+    "react-hot-loader": "^1.2.8",
+    "sinon": "^1.15.4",
+    "sinon-chai": "^2.8.0",
+    "webpack-dev-server": "^1.10.0"
+  }
 }
