--- src/editor.jsx
@@ -1,43 +1,0 @@
-/* eslint new-cap:0 no-unused-vars:0 */
-"use strict";
-
-import React from "react";
-
-const Editor = React.createClass({
-  componentDidMount() {
-    this.editor = CodeMirror.fromTextArea(this.refs.editor.getDOMNode(), {
-      mode: "javascript",
-      lineNumbers: false,
-      lineWrapping: true,
-      smartIndent: false,
-      matchBrackets: true,
-      theme: this.props.theme,
-      readOnly: this.props.readOnly
-    });
-    this.editor.on("change", this._handleChange);
-  },
-
-  componentDidUpdate() {
-    if (this.props.readOnly) {
-      this.editor.setValue(this.props.codeText);
-    }
-  },
-
-  _handleChange() {
-    if (!this.props.readOnly && this.props.onChange) {
-      this.props.onChange(this.editor.getValue());
-    }
-  },
-
-  render() {
-    var editor = <textarea ref="editor" defaultValue={this.props.codeText} />;
-
-    return (
-      <div style={this.props.style} className={this.props.className}>
-        {editor}
-      </div>
-    );
-  }
-});
-
-export default Editor;
