+++ test/client/test.html
@@ -1,0 +1,18 @@
+<!DOCTYPE HTML>
+<html>
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <title>Frontend Tests</title>
+    <link rel="stylesheet" type="text/css" href="/node_modules/mocha/mocha.css">
+  </head>
+  <body>
+    <div id="mocha" />
+
+    <!-- JavaScript Test Libraries. -->
+    <script type="text/javascript" src="/node_modules/mocha/mocha.js"></script>
+    <script type="text/javascript" src="/node_modules/sinon/pkg/sinon.js"></script>
+
+    <!-- JavaScript Test Spec. -->
+    <script src="http://127.0.0.1:3001/assets/main.js"></script>
+  </body>
+</html>
